wip TestSingleAssemblyDeploymentUnit
